requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1
dj-database-url==0.5.0
Django==2.2
django-crontab==0.7.1
django-filter==2.3.0
django-heroku==0.3.1
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-scheduler==0.9.3
djangorestframework==3.11.1
gunicorn==20.0.4
heroku==0.1.4
icalendar==4.0.7
idna==2.9
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2==2.8.5
pycparser==2.20
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
schedule==0.6.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.0.1
zipp==3.1.0
